Finish cleaning up imagestack tests.
1. Rename test_imagestack_*.py to test_*.py
2. Break up test_imagestack.py, which became a kitchen sink.

Test plan: `make fast-test`
