+++ starfish/imagestack/test/test_coordinates.py
