+++ starfish/imagestack/test/test_cropped_load.py
