+++ starfish/imagestack/test/test_index.py
