+++ starfish/imagestack/test/test_labeled_indices.py
