+++ starfish/imagestack/test/test_metadata.py
