+++ starfish/imagestack/test/test_slice_ops.py
@@ -1,23 +1,7 @@
-from collections import OrderedDict
-
 import numpy as np
-import pytest
-import xarray as xr
-from slicedimage import ImageFormat
 
-from starfish.imagestack.imagestack import ImageStack
-from starfish.imagestack.physical_coordinate_calculator import get_physical_coordinates_of_z_plane
-from starfish.intensity_table.test.factories import synthetic_intensity_table
-from starfish.test.factories import (
-    synthetic_spot_pass_through_stack,
-    SyntheticData)
-from starfish.types import Axes, PhysicalCoordinateTypes
-from .factories import imagestack_with_coords_factory, synthetic_stack
-from .imagestack_test_utils import verify_physical_coordinates
-
-X_COORDS = 1, 2
-Y_COORDS = 4, 6
-Z_COORDS = 1, 3
+from starfish.types import Axes
+from .factories import synthetic_stack
 
 
 def test_get_slice_simple_index():
@@ -175,108 +159,3 @@ def test_set_slice_range():
     stack.set_slice(index, expected, [Axes.ROUND, Axes.CH, Axes.ZPLANE])
 
     assert np.array_equal(stack.get_slice(index)[0], expected)
-
-
-def test_from_numpy_array_raises_error_when_incorrect_dims_passed():
-    array = np.ones((2, 2), dtype=np.float32)
-    # verify this method works with the correct shape
-    image = ImageStack.from_numpy_array(array.reshape((1, 1, 1, 2, 2)))
-    assert isinstance(image, ImageStack)
-
-    with pytest.raises(ValueError):
-        ImageStack.from_numpy_array(array.reshape((1, 1, 2, 2)))
-        ImageStack.from_numpy_array(array.reshape((1, 2, 2)))
-        ImageStack.from_numpy_array(array)
-        ImageStack.from_numpy_array(array.reshape((1, 1, 1, 1, 2, 2)))
-
-
-def test_from_numpy_array_automatically_handles_float_conversions():
-    x = np.zeros((1, 1, 1, 20, 20), dtype=np.uint16)
-    stack = ImageStack.from_numpy_array(x)
-    assert stack.xarray.dtype == np.float32
-
-
-def test_max_projection_preserves_dtype():
-    original_dtype = np.float32
-    array = np.ones((2, 2, 2), dtype=original_dtype)
-    image = ImageStack.from_numpy_array(array.reshape((1, 1, 2, 2, 2)))
-
-    max_projection = image.max_proj(Axes.CH, Axes.ROUND, Axes.ZPLANE)
-    assert max_projection.xarray.dtype == original_dtype
-
-
-def test_synthetic_spot_creation_raises_error_with_coords_too_small():
-    num_z = 0
-    height = 40
-    width = 50
-    intensity_table = synthetic_intensity_table()
-    with pytest.raises(ValueError):
-        SyntheticData.synthetic_spots(intensity_table, num_z, height, width)
-
-
-def test_synthetic_spot_creation_produces_an_imagestack_with_correct_spot_location():
-
-    codebook, true_intensities, image = synthetic_spot_pass_through_stack()
-
-    g, c, r = np.where(true_intensities.values)
-
-    x = np.empty_like(g)
-    y = np.empty_like(g)
-    z = np.empty_like(g)
-    breaks = np.concatenate([
-        np.array([0]),
-        np.where(np.diff(g))[0] + 1,
-        np.array([g.shape[0]])
-    ])
-    for i in np.arange(len(breaks) - 1):
-        x[breaks[i]: breaks[i + 1]] = true_intensities.coords[Axes.X.value][i]
-        y[breaks[i]: breaks[i + 1]] = true_intensities.coords[Axes.Y.value][i]
-        z[breaks[i]: breaks[i + 1]] = true_intensities.coords[Axes.ZPLANE.value][i]
-
-    # only 8 values should be set, since there are only 8 locations across the tensor
-    assert np.sum(image.xarray != 0) == 8
-
-    intensities = image.xarray.sel(
-        x=xr.DataArray(x, dims=['intensity']),
-        y=xr.DataArray(y, dims=['intensity']),
-        z=xr.DataArray(z, dims=['intensity']),
-        r=xr.DataArray(r, dims=['intensity']),
-        c=xr.DataArray(c, dims=['intensity']))
-    assert np.allclose(
-        intensities,
-        true_intensities.values[np.where(true_intensities)])
-
-
-@pytest.mark.parametrize("format,count", (
-    (ImageFormat.TIFF, 6),
-    (ImageFormat.NUMPY, 6),
-))
-def test_imagestack_export(tmpdir, format, count, recwarn):
-    """
-    Save a synthetic stack to files and check the results
-    """
-    stack_shape = OrderedDict([(Axes.ROUND, 3), (Axes.CH, 2),
-                               (Axes.ZPLANE, 1), (Axes.Y, 50), (Axes.X, 40)])
-
-    physical_coords = OrderedDict([(PhysicalCoordinateTypes.X_MIN, X_COORDS[0]),
-                                   (PhysicalCoordinateTypes.X_MAX, X_COORDS[1]),
-                                   (PhysicalCoordinateTypes.Y_MIN, Y_COORDS[0]),
-                                   (PhysicalCoordinateTypes.Y_MAX, Y_COORDS[1]),
-                                   (PhysicalCoordinateTypes.Z_MIN, Z_COORDS[0]),
-                                   (PhysicalCoordinateTypes.Z_MAX, Z_COORDS[1])])
-
-    stack = imagestack_with_coords_factory(stack_shape, physical_coords)
-
-    stack_json = tmpdir / "output.json"
-    stack.export(str(stack_json), tile_format=format)
-    files = list([x for x in tmpdir.listdir() if str(x).endswith(format.file_ext)])
-    loaded_stack = ImageStack.from_path_or_url(str(stack_json))
-    verify_physical_coordinates(
-        loaded_stack,
-        X_COORDS,
-        Y_COORDS,
-        get_physical_coordinates_of_z_plane(Z_COORDS),
-    )
-    assert count == len(files)
-    with open(files[0], "rb") as fh:
-        format.reader_func(fh)
